[collector]feature:add CustomizedJmx feature
